+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Roles/List/Validator.cs
@@ -1,0 +1,7 @@
+using FluentValidation;
+
+namespace TimeSheet.Modules.EmploymentManagement.Application.Roles.List;
+
+public sealed class Validator : AbstractValidator<Query>
+{
+}